requirements.txt
Appium-Python-Client==2.2.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
behave==1.2.6
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
cryptography==37.0.0
glob2==0.7
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
Mako==1.2.0
MarkupSafe==2.1.1
outcome==1.1.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
wsproto==1.1.0
zipp==3.8.0
